Fixed file error issue

+++ pages/page_1.py
@@ -20,12 +20,17 @@ def image_to_base64(path):
 
 def main():
 
-    try: 
-        st.write(os.listdir(Path.home() / ".config" / "localStoragePy"))
-    
-    except FileNotFoundError:
+    file = Path.home() / ".config" / "localStoragePy"
+
+    if not file.exists():
         with open("last_cleanup.txt", "w") as f:
             f.write(str(datetime.now().isoformat()))
+    elif path.exists():
+        content = f.read_text().strip()
+
+        if not content:
+            with open("last_cleanup.txt", "w") as f:
+                f.write(str(datetime.now().isoformat()))
 
 
     with open("last_cleanup.txt", "r") as f:
